+++ database/migrations/2023_03_30_171233_create_class_detail_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('class_detail', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
+            $table->double('price');
             $table->timestamps();
         });
     }
